requirements.txt
aiocoap==0.3
aiohttp==1.0.5
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
dataclasses==0.8
docopt==0.6.2
esmre==0.5.2
et-xmlfile==1.1.0
idna==3.3
importlib-resources==5.4.0
Js2Py==0.71
libs==0.0.10
lxml==4.6.4
multidict==5.2.0
numpy==1.19.5
openpyxl==3.0.9
pandas==1.1.5
pipreqs==0.4.11
PyExecJS==1.5.1
pyjsparser==2.7.1
pymongo==3.5.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
selenium==3.141.0
sites==0.0.1
six==1.16.0
tenacity==8.0.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
zipp==3.6.0
